osteolab-ml-platform/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-learn
joblib
pytest
mlflow>=3.10.0
pyyaml
psycopg2-binary